server/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pymongo==4.5.0
python-dotenv==1.0.0
pillow==10.0.0
pydantic==2.4.2
tensorflow==2.13.0
numpy==1.24.3
tensorflow-hub==0.14.0
opencv-python==4.8.0.76 